+++ noodles-fasta/src/fai/async.rs
@@ -1,3 +1,4 @@
 //! Async FAI.
 
+pub mod fs;
 pub mod io;